pyproject.toml
[project]
name = "contract-deidentification"
version = "0.1.0"
description = "合同文本脱敏工具，支持四层脱敏策略"
readme = "README.md"
requires-python = ">=3.9,<=3.11"
license = {text = "MIT"}
authors = [
    {name = "Contract Deidentification Team"}
]
keywords = ["deidentification", "privacy", "contract", "anonymization", "ner"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "faker>=20.0.0",
    "modelscope>=1.9.0",  # 推荐使用 ModelScope 下载模型
    "paddlenlp>=2.6.0",  # 向后兼容，可选
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",  # 用于读取 .env 文件
    "tool-helpers>=0.1.2",  # Override to get ARM64 support for Python 3.10
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
contract-deid = "contract_deid.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contract_deid"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
